+++ pyfragment/methods/__init__.py
@@ -4,6 +4,7 @@ from pyfragment.methods.giveaway_stars import giveaway_stars
 from pyfragment.methods.purchase_premium import purchase_premium
 from pyfragment.methods.purchase_stars import purchase_stars
 from pyfragment.methods.recharge_ads import recharge_ads
+from pyfragment.methods.search_auctions import search_auctions
 from pyfragment.methods.topup_ton import topup_ton
 
 __all__ = [
@@ -13,6 +14,7 @@ __all__ = [
     "purchase_premium",
     "purchase_stars",
     "recharge_ads",
+    "search_auctions",
     "terminate_sessions",
     "toggle_login_codes",
     "topup_ton",