+++ pyfragment/utils/__init__.py
@@ -1,5 +1,5 @@
 from pyfragment.utils.decoder import clean_decode
-from pyfragment.utils.html import parse_login_code
+from pyfragment.utils.html import parse_auction_rows, parse_login_code
 from pyfragment.utils.http import (
     execute_transaction_request,
     fragment_request,
@@ -11,6 +11,7 @@ from pyfragment.utils.wallet import get_account_info, process_transaction
 
 __all__ = [
     "clean_decode",
+    "parse_auction_rows",
     "parse_login_code",
     "execute_transaction_request",
     "fragment_request",